// gen/tasks_pb.ts
export class Stats extends Message<Stats> {
  /**
   * Timestamp of when the task was spawned.
   *
   * @generated from field: google.protobuf.Timestamp created_at = 1;
   */
  createdAt?: Timestamp;

  /**
   * Timestamp of when the task was dropped.
   *
   * @generated from field: google.protobuf.Timestamp dropped_at = 2;
   */
  droppedAt?: Timestamp;

  /**
   * The total number of times this task has been woken over its lifetime.
   *
   * @generated from field: uint64 wakes = 3;
   */
  wakes = protoInt64.zero;

  /**
   * The total number of times this task's waker has been cloned.
   *
   * @generated from field: uint64 waker_clones = 4;
   */
  wakerClones = protoInt64.zero;

  /**
   * The total number of times this task's waker has been dropped.
   *
   * @generated from field: uint64 waker_drops = 5;
   */
  wakerDrops = protoInt64.zero;

  /**
   * The timestamp of the most recent time this task has been woken.
   *
   * If this is `None`, the task has not yet been woken.
   *
   * @generated from field: optional google.protobuf.Timestamp last_wake = 6;
   */
  lastWake?: Timestamp;

  /**
   * Contains task poll statistics.
   *
   * @generated from field: rs.tokio.console.common.PollStats poll_stats = 7;
   */
  pollStats?: PollStats;

  /**
   * The total number of times this task has woken itself.
   *
   * @generated from field: uint64 self_wakes = 8;
   */
  selfWakes = protoInt64.zero;

  /**
   * The total duration this task was scheduled prior to being polled, summed
   * across all poll cycles.
   *
   * Note that this includes only polls that have started, and does not
   * reflect any scheduled state where the task hasn't yet been polled.
   * Subtracting both `busy_time` (from the task's `PollStats`) and
   * `scheduled_time` from the total lifetime of the task results in the
   * amount of time it spent unable to progress because it was waiting on 
   * some resource.
   *
   * @generated from field: google.protobuf.Duration scheduled_time = 9;
   */
  scheduledTime?: Duration;

  constructor(data?: PartialMessage<Stats>) {
    super();
    proto3.util.initPartial(data, this);
  }

  static readonly runtime: typeof proto3 = proto3;
  static readonly typeName = "rs.tokio.console.tasks.Stats";
  static readonly fields: FieldList = proto3.util.newFieldList(() => [
    { no: 1, name: "created_at", kind: "message", T: Timestamp },
    { no: 2, name: "dropped_at", kind: "message", T: Timestamp },
    { no: 3, name: "wakes", kind: "scalar", T: 4 /* ScalarType.UINT64 */ },
    { no: 4, name: "waker_clones", kind: "scalar", T: 4 /* ScalarType.UINT64 */ },
    { no: 5, name: "waker_drops", kind: "scalar", T: 4 /* ScalarType.UINT64 */ },
    { no: 6, name: "last_wake", kind: "message", T: Timestamp, opt: true },
    { no: 7, name: "poll_stats", kind: "message", T: PollStats },
    { no: 8, name: "self_wakes", kind: "scalar", T: 4 /* ScalarType.UINT64 */ },
    { no: 9, name: "scheduled_time", kind: "message", T: Duration },
  ]);

  static fromBinary(bytes: Uint8Array, options?: Partial<BinaryReadOptions>): Stats {
    return new Stats().fromBinary(bytes, options);
  }

  static fromJson(jsonValue: JsonValue, options?: Partial<JsonReadOptions>): Stats {
    return new Stats().fromJson(jsonValue, options);
  }

  static fromJsonString(jsonString: string, options?: Partial<JsonReadOptions>): Stats {
    return new Stats().fromJsonString(jsonString, options);
  }

  static equals(a: Stats | PlainMessage<Stats> | undefined, b: Stats | PlainMessage<Stats> | undefined): boolean {
    return proto3.util.equals(Stats, a, b);
  }
}